Proper arangement of errors and cli helpers
