+++ lib/crowbar/client/cli.rb
@@ -14,9 +14,13 @@
 # limitations under the License.
 #
 
-require_relative "cli/helper"
-require_relative "cli/flags"
-require_relative "cli/hooks"
+require "gli"
+
+GLI::Commands::Help.tap do |config|
+  config.skips_around = false
+  config.skips_pre = false
+  config.skips_post = false
+end
 
 module Crowbar
   module Client
@@ -33,9 +37,52 @@ module Crowbar
         value.split(/,/).map(&:strip)
       end
 
+      desc "Path to a configuration file"
+      flag [:c, :config], default_value: nil
+
+      desc "Alias for a config section"
+      flag [:a, :alias], default_value: "default"
+
+      desc "Specify username for connection"
+      flag [:U, :username], default_value: "crowbar"
+
+      desc "Specify password for connection"
+      flag [:P, :password], default_value: "crowbar"
+
+      desc "Specify host for connection"
+      flag [:n, :hostname], default_value: "http://127.0.0.1"
+
+      desc "Specify port for connection"
+      flag [:p, :port], default_value: "80"
+
+      desc "Specify timeout for connection"
+      flag [:t, :timeout], default_value: "60"
+
+      desc "Output debug informations"
+      switch [:d, :debug], negatable: false
+
+      desc "Call legacy routes for 1.x compatibility"
+      switch [:l, :legacy], negatable: false
+
+      pre do |global|
+        ENV["GLI_DEBUG"] = "true" if global[:debug]
+
+        config = configure(global[:config], global[:alias])
+
+        Request::Party.instance.configure(
+          host:     config[:hostname] || global[:hostname],
+          port:     config[:port] || global[:port],
+          username: config[:username] || global[:username],
+          password: config[:password] || global[:password],
+          legacy:   config[:legacy] || global[:legacy],
+          debug:    config[:debug] || global[:debug],
+          timeout:  config[:timeout] || global[:timeout]
+        )
+
+        true
+      end
+
       include Cli::Helper
-      include Cli::Flags
-      include Cli::Hooks
 
       include Command::Barclamps
       include Command::Batch
