--- lib/crowbar/client/errors.rb
@@ -1,43 +1,0 @@
-#
-# Copyright 2015, SUSE Linux GmbH
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-module Crowbar
-  module Client
-    class UnavailableBarclampError < StandardError
-      def initialize(barclamp)
-        super("Barclamp #{barclamp} is not available")
-      end
-    end
-
-    class BadFormatterError < StandardError
-    end
-
-    class BadFilterError < StandardError
-    end
-
-    class InvalidFormatError < StandardError
-    end
-
-    class EditorAbortError < StandardError
-    end
-
-    class EditorStartupError < StandardError
-    end
-
-    class InvalidJsonError < StandardError
-    end
-  end
-end
